+++ server/routes/manual-data.js
@@ -3,7 +3,7 @@
 const express = require('express');
 const router = express.Router();
 
-const { requireScopeForApiRoute } = require('../auth/authorization');
+const { requireScopeForApiRoute } = require('@cotarr/collab-backend-token-auth');
 const controller = require('../controllers/manual-data');
 const validations = require('../validations/manual-data');